+++ src/WebformBuilder.js
@@ -3,7 +3,7 @@ import dragula from 'dragula';
 import Tooltip from 'tooltip.js';
 import Components from './components/Components';
 import BuilderUtils from './utils/builder';
-import { getComponent, bootstrapVersion } from './utils/utils';
+import { getComponent, bootstrapVersion, eachComponent } from './utils/utils';
 import EventEmitter from 'eventemitter2';
 import Promise from 'native-promise-only';
 import _ from 'lodash';
@@ -183,6 +183,25 @@ export default class WebformBuilder extends Webform {
   }
 
   setForm(form) {
+    //populate isEnabled for recaptcha form settings
+    var isRecaptchaEnabled = false;
+    if (form.components) {
+      eachComponent(form.components, component => {
+        if (isRecaptchaEnabled) {
+          return;
+        }
+        if (component.type === 'recaptcha') {
+          isRecaptchaEnabled = true;
+          return false;
+        }
+      });
+      if (isRecaptchaEnabled) {
+        _.set(form, 'settings.recaptcha.isEnabled', true);
+      }
+      else if (_.get(form, 'settings.recaptcha.isEnabled')) {
+        _.set(form, 'settings.recaptcha.isEnabled', false);
+      }
+    }
     this.emit('change', form);
     return super.setForm(form).then(retVal => {
       setTimeout(() => (this.builderHeight = this.element.offsetHeight), 200);